feature: add interfaces for extension managements & translation hooks
The Manager interface handles and maintains the registered extensions. It contains methods to determine whether an API group is supported by a registered extension and methods to return hook clients for different hook types (e.g. XDSHook)

Each type of translation hook has its own interface that contains methods representing different types of hook entrypoints for that particular translation step.

Finally, we some of the proto objects that the Envoy Gateway Control Plane and the extension service will be interacting with. These objects provide the context needed for the extension to modify the resource according to user intent. The actual gRPC service spec will be specified separately.

Signed-off-by: Hamzah Qudsi <hqudsi@datawire.io>